Move dataplane/server package -> dataplane
server is too generic of a name.
